Extract Aggregator base class
